fix the bug of convertedDstIp
this field should be 32 bits.

unsigned long in 32-bit system will be 32 bits, but in 64-bit system will be 64 bits.

when running this kernel tap project in 64-bit system, it may occurs some error.

+++ kudp.c
@@ -16,7 +16,7 @@ struct my_kudp
     char dstip[IPADDR_SIZE];
     int  tunnelport;
 
-    unsigned long convertedDstip;
+    unsigned int convertedDstip;
 
     struct socket* sock;
